Factor `struct iked` initialization out
